Updated headers (license)

+++ src/common.hpp
@@ -1,7 +1,24 @@
+/*
+ * MPIRPC: MPI based invocation of functions on other ranks
+ * Copyright (C) 2014  Colin MacLean <s0838159@sms.ed.ac.uk>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
 #ifndef COMMON_HPP
 #define COMMON_HPP
-
-//#include <QtGlobal>
 
 #ifndef NDEBUG
 #   define ASSERT(condition, message) \
@@ -45,9 +62,9 @@ struct FunctionParts<R(*)(Args...)>
     using function_type = R(*)(Args...);
 };
 
-using FunctionHandle = unsigned long;
-using TypeId = unsigned long;
-using ObjectId = unsigned long;
+using FunctionHandle = unsigned long long;
+using TypeId = unsigned long long;
+using ObjectId = unsigned long long;
 
 }
